requirements.txt
langchain>=0.2.0
langchain-community>=0.2.0
langchain-groq>=0.1.3
faiss-cpu>=1.7.4
sentence-transformers>=2.7.0
requests>=2.31.0
python-dotenv>=1.0.0
streamlit>=1.32.0